backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
django-filter==23.3
djoser==2.1.0
gunicorn==20.1.0
Pillow==10.1.0
psycopg2-binary==2.9.3
reportlab==4.0.6